requirements.txt
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
shap>=0.41.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
gitpython>=3.1.0
tree-sitter>=0.20.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
networkx>=2.8.0
PyYAML>=6.0
rouge-score>=0.1.2
nltk>=3.7
tqdm>=4.64.0
joblib>=1.1.0